+++ src/models/AclRole.js
@@ -4,17 +4,13 @@ const mongoose = require('mongoose');
 
 const { Schema } = mongoose;
 
-async function aclRoleModel(uw) {
-  const schema = new Schema({
-    _id: String,
-    roles: [{ type: String, ref: 'AclRole', index: true }],
-  }, {
-    collection: 'acl_roles',
-    idKey: 'name',
-    minimize: true,
-  });
+const schema = new Schema({
+  _id: String,
+  roles: [{ type: String, ref: 'AclRole', index: true }],
+}, {
+  collection: 'acl_roles',
+  idKey: 'name',
+  minimize: true,
+});
 
-  uw.mongo.model('AclRole', schema);
-}
-
-module.exports = aclRoleModel;
+module.exports = schema;