+++ src/routes/search.js
@@ -3,7 +3,9 @@
 const router = require('router');
 const route = require('../route');
 const protect = require('../middleware/protect');
+const schema = require('../middleware/schema');
 const controller = require('../controllers/search');
+const validations = require('../validations');
 
 function searchRoutes() {
   return router()
@@ -11,11 +13,13 @@ function searchRoutes() {
     // GET /search/ - Search for media across all sources.
     .get(
       '/',
+      schema(validations.searchAll),
       route(controller.searchAll),
     )
     // GET /search/:source - Search for media in a single source.
     .get(
       '/:source',
+      schema(validations.search),
       route(controller.search),
     );
 }