+++ src/validations.js
@@ -332,10 +332,45 @@ exports.addPlaylistItems = {
     properties: {
       items: {
         type: 'array',
-        items: { $ref: 'https://ns.u-wave.net/schemas/definitions.json#/definitions/ObjectID' },
+        items: {
+          type: 'object',
+          properties: {
+            sourceType: { type: 'string' },
+            sourceID: {
+              oneOf: [{ type: 'string' }, { type: 'number' }],
+            },
+            artist: { type: 'string' },
+            title: { type: 'string' },
+          },
+          required: ['sourceType', 'sourceID', 'artist', 'title'],
+        },
       },
     },
     required: ['items'],
+
+    // Different ways to describe the insert position
+    oneOf: [
+      {
+        type: 'object',
+        properties: {
+          after: {
+            oneOf: [
+              { $ref: 'https://ns.u-wave.net/schemas/definitions.json#/definitions/ObjectID' },
+              { const: null },
+              { const: -1 },
+            ],
+          },
+        },
+        required: ['after'],
+      },
+      {
+        type: 'object',
+        properties: {
+          at: { enum: ['start', 'end'] },
+        },
+        required: ['at'],
+      },
+    ],
   },
 };
 
@@ -364,6 +399,7 @@ exports.movePlaylistItems = {
       },
     },
     required: ['items'],
+    // Different ways to describe the insert position
     oneOf: [
       {
         type: 'object',
@@ -411,6 +447,35 @@ exports.updatePlaylistItem = {
 
 exports.removePlaylistItem = {
   params: playlistItemParams,
+};
+
+// Validations for search routes:
+
+exports.searchAll = {
+  query: {
+    type: 'object',
+    properties: {
+      query: { type: 'string' },
+    },
+    required: ['query'],
+  },
+};
+
+exports.search = {
+  query: {
+    type: 'object',
+    properties: {
+      query: { type: 'string' },
+    },
+    required: ['query'],
+  },
+  params: {
+    type: 'object',
+    properties: {
+      source: { type: 'string', minLength: 1 },
+    },
+    required: ['source'],
+  },
 };
 
 // Validations for user routes: